backend/requirements.txt
autopep8
django
django-environ
django-extensions
djangorestframework
psycopg2-binary
django-admin-interface
